@@ -3,13 +3,15 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.alicebot.ab</groupId>
 	<artifactId>ab</artifactId>
-	<version>4.5.1_SPRINT_FIX</version>
+	<version>4.6.1-sprint</version>
 	<name>sprint-alicebot</name>
 	<description>
 		Program AB is the reference implementation of the AIML 2.0 draft specification.
-		AIML is a widely adopted standard for creating chat bots and mobile virtual assistants like ALICE, Mitsuku, English Tutor, The Professor, S.U.P.E.R. and many more.</description>
+		AIML is a widely adopted standard for creating chat bots and mobile virtual assistants like ALICE, Mitsuku, English Tutor, The Professor, S.U.P.E.R. and many more.
+	</description>
 	<properties>
-		<slf4j.version>1.7.36</slf4j.version>
+		<slf4j.version>2.0.13</slf4j.version>
+		<java.version>21</java.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 	</properties>
@@ -49,67 +51,50 @@
 		<dependency>
 			<groupId>pl.sprint</groupId>
 			<artifactId>validator</artifactId>
-			<version>1.5.6_FIX</version>
+			<version>2.0.4</version>
 		</dependency>
-
-		<!-- ML -->
+		<!-- ML FastText Libraries -->
 		<dependency>
 			<groupId>com.mayabot.mynlp</groupId>
-			<artifactId>fastText4j</artifactId>
-			<version>3.1.7</version>
+			<artifactId>mynlp</artifactId>
+			<version>4.1.4-sprint</version>
 		</dependency>
+		<!-- ML Libraries -->
 
-		<!-- MLA -->
-		<!-- https://github.com/GitHubSprint/fasttext4j -->
-		<dependency>
-			<groupId>fasttext</groupId>
-			<artifactId>fasttext4j</artifactId>
-			<version>0.2.3_SPRINT</version>
-		</dependency>
 	</dependencies>
 	<build>
-            <plugins>
-                <plugin>
-                    <artifactId>maven-compiler-plugin</artifactId>
-                    <version>3.5.1</version>
-                    <configuration>
-                            <source>1.8</source>
-                            <target>1.8</target>
-                    </configuration>
-                </plugin>
-                <plugin>
-                    <artifactId>maven-assembly-plugin</artifactId>
-                    <configuration>
-                        <archive>
-                            <manifest>
-                              <mainClass>org.alicebot.ab.cli.Main</mainClass>
-                            </manifest> 
-                        </archive>
-                      <descriptorRefs>
-                        <descriptorRef>jar-with-dependencies</descriptorRef>
-                      </descriptorRefs>
-                    </configuration>
-                    <executions>
-                      <execution>
-                        <id>simple-command</id>
-                        <phase>package</phase>
-                        <goals>
-                          <goal>single</goal>
-                        </goals>
-                      </execution>
-                    </executions>
-                  </plugin>
-            </plugins>
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.1</version>
+				<configuration>
+					<source>21</source>
+					<target>21</target>
+				</configuration>
+			</plugin>
+			<plugin>
+				<artifactId>maven-assembly-plugin</artifactId>
+				<configuration>
+					<archive>
+						<manifest>
+						  <mainClass>org.alicebot.ab.cli.Main</mainClass>
+						</manifest>
+					</archive>
+				  <descriptorRefs>
+					<descriptorRef>jar-with-dependencies</descriptorRef>
+				  </descriptorRefs>
+				</configuration>
+				<executions>
+				  <execution>
+					<id>simple-command</id>
+					<phase>package</phase>
+					<goals>
+					  <goal>single</goal>
+					</goals>
+				  </execution>
+				</executions>
+			  </plugin>
+		</plugins>
 	</build>
-	<profiles>
-		<profile>
-			<id>disable-java8-doclint</id>
-			<activation>
-				<jdk>[1.8,)</jdk>
-			</activation>
-			<properties>
-				<additionalparam>-Xdoclint:none</additionalparam>
-			</properties>
-		</profile>
-	</profiles>
 </project>
