@@ -35,65 +35,21 @@ import org.w3c.dom.NodeList;
  * This is just a stub to make the contactaction.aiml model work on a PC
  with some extension tags that are defined for mobile devices.
  */
-public class PCAIMLProcessorExtension implements AIMLProcessorExtension 
-{
+public class PCAIMLProcessorExtension implements AIMLProcessorExtension {
     private static final Logger log = LoggerFactory.getLogger(PCAIMLProcessorExtension.class);
     
     public Set<String> extensionTagNames = Utilities.stringSet("contactid","multipleids","displayname","dialnumber","emailaddress","contactbirthday","addinfo");
-
-    static 
-    {
-        SprintUtils.mlaModels = getMLAModels();        
-    } 
     
-    static 
+    static
     {
         SprintUtils.mlModels = getMLModels(); 
     } 
-    
-    
-    public static Map<String,fasttext.FastText> getMLAModels() 
-    {
-        Map<String, fasttext.FastText> mlaModels = new HashMap<String, fasttext.FastText>();    
-        try 
-        {
-            String path = new File(".").getCanonicalPath().replace("\\", "/") + "/models/";   
 
-            // Directory path here
-            String file;
-            String filePath;
-            File folder = new File(path);
-            if (folder.exists()) {
-                File[] listOfFiles = folder.listFiles();
-                log.info("Loading MLA Model files from '{}'", path);
-                for (File listOfFile : listOfFiles) {
-                    if (listOfFile.isFile()) {
-                        file = listOfFile.getName();
-                        filePath = listOfFile.getAbsoluteFile().toString();
-                        if (file.endsWith(".ftz") || file.endsWith(".FTZ")) {
-                            log.info("Read MLA Model file: " + file + " filePath: " + filePath);
-                            String modelName = file.substring(0, file.length()-".ftz".length());
-                            log.info("Read MLA Model "+modelName);
-                            fasttext.FastText ftmodel = fasttext.FastText.loadModel(filePath);
 
-                            mlaModels.put(modelName, ftmodel);
-                        }
-                    }
-                }
-            }
-            else log.info("addMLAModels: '{}' does not exist.", path);
-        } catch (Exception ex)  {
-            ex.printStackTrace();
-        }
-        
-        return mlaModels;        
-    }
+    public static Map<String, FastText> getMLModels() {
+        Map<String, FastText> mlModels = new HashMap<>();
 
-    public static Map<String, FastText> getMLModels()
-    {
-        Map<String, FastText> mlModels = new HashMap<String, FastText>();
-        try
-        {
+        try {
             String path = new File(".").getCanonicalPath().replace("\\", "/") + "/models/";
 
             // Directory path here
@@ -102,67 +58,33 @@ public class PCAIMLProcessorExtension implements AIMLProcessorExtension
             File folder = new File(path);
             if (folder.exists()) {
                 File[] listOfFiles = folder.listFiles();
+                if(listOfFiles == null)
+                    return mlModels;
+
                 log.info("Loading ML Model files from '{}'", path);
                 for (File listOfFile : listOfFiles) {
                     if (listOfFile.isFile()) {
                         file = listOfFile.getName();
                         filePath = listOfFile.getAbsoluteFile().toString();
                         if (file.endsWith(".bin") || file.endsWith(".BIN")) {
-                            log.info("Read ML Model file: " + file + " filePath: " + filePath);
+                            log.info("Read ML Model file: {} filePath: {}", file, filePath);
                             String modelName = file.substring(0, file.length()-".bin".length());
-                            log.info("Read ML Model "+modelName);
+                            log.info("Read ML Model {}", modelName);
                             FastText fastText = FastText.Companion.loadModelFromSingleFile(new File(filePath));
                             mlModels.put(modelName, fastText);
                         }
                     }
                 }
             }
-            else log.info("addMLAodels: '{}' does not exist.", path);
+            else
+                log.info("getMLModels path: '{}' does not exist.", path);
         } catch (Exception ex)  {
-            ex.printStackTrace();
+            log.error("getMLModels Error", ex);
         }
 
         return mlModels;
     }
-    
-//    public static Map<String, FastText> getMLModels()
-//    {
-//        Map<String, FastText> mlModels = new HashMap<String, FastText>();
-//        try
-//        {
-//            String path = new File(".").getCanonicalPath().replace("\\", "/") + "/models/";
-//
-//            // Directory path here
-//            String model;
-//            String modelPath;
-//            File folder = new File(path);
-//            if (folder.exists()) {
-//                File[] listOfFiles = folder.listFiles();
-//                log.info("Loading ML Model files from '{}'", path);
-//                for (File listOfFile : listOfFiles) {
-//                    if (listOfFile.isDirectory()) {
-//                        model = listOfFile.getName();
-//                        modelPath = listOfFile.getAbsoluteFile().toString();
-//
-//                        log.info("Read ML Model folder: " + model + " filePath: " + modelPath);
-//
-//                        FastText fastText = FastText.Companion.loadModel(new File(modelPath), true);
-//                        //fasttext.FastText ftmodel = fasttext.FastText.loadModel(modelPath, true);
-//
-//
-//                        mlModels.put(model, fastText);
-//
-//                    }
-//                }
-//            }
-//            else log.info("addMLModels: '{}' does not exist.", path);
-//        } catch (Exception ex)  {
-//            ex.printStackTrace();
-//        }
-//
-//        return mlModels;
-//    }
-    
+
     
     
     @Override
@@ -203,21 +125,15 @@ public class PCAIMLProcessorExtension implements AIMLProcessorExtension
     }
     private String contactId(Node node, ParseState ps) {
         String displayName = AIMLProcessor.evalTagContent(node, ps, null);
-        String result = Contact.contactId(displayName);
-        //log.info("contactId("+displayName+")="+result);
-        return result;
+        return Contact.contactId(displayName);
     }
     private String multipleIds(Node node, ParseState ps){
         String contactName = AIMLProcessor.evalTagContent(node, ps, null);
-        String result = Contact.multipleIds(contactName);
-        //log.info("multipleIds("+contactName+")="+result);
-        return result;
+        return Contact.multipleIds(contactName);
     }
     private String displayName(Node node, ParseState ps){
         String id = AIMLProcessor.evalTagContent(node, ps, null);
-        String result = Contact.displayName(id);
-        //log.info("displayName("+id+")="+result);
-        return result;
+        return Contact.displayName(id);
     }
     private String dialNumber(Node node, ParseState ps) {
         NodeList childList = node.getChildNodes();
@@ -231,9 +147,7 @@ public class PCAIMLProcessorExtension implements AIMLProcessorExtension
                 type = AIMLProcessor.evalTagContent(childList.item(i), ps, null);
             }
         }
-        String result = Contact.dialNumber(type, id);
-        //log.info("dialNumber("+id+")="+result);
-        return result;
+        return Contact.dialNumber(type, id);
     }
 
     private String emailAddress(Node node, ParseState ps){
@@ -248,42 +162,30 @@ public class PCAIMLProcessorExtension implements AIMLProcessorExtension
                 type = AIMLProcessor.evalTagContent(childList.item(i), ps, null);
             }
         }
-        String result = Contact.emailAddress(type, id);
-        //log.info("emailAddress("+id+")="+result);
-        return result;
+        return Contact.emailAddress(type, id);
     }
 
 
     private String contactBirthday(Node node, ParseState ps){
         String id = AIMLProcessor.evalTagContent(node, ps, null);
-        String result = Contact.birthday(id);
-        //log.info("birthday("+id+")="+result);
-        return result;
+        return Contact.birthday(id);
     }
     @Override
 	public String recursEval(Node node, ParseState ps) {
         try {
             String nodeName = node.getNodeName();
-            switch (nodeName) {
-                case "contactid":
-                    return contactId(node, ps);
-                case "multipleids":
-                    return multipleIds(node, ps);
-                case "dialnumber":
-                    return dialNumber(node, ps);
-                case "addinfo":
-                    return newContact(node, ps);
-                case "displayname":
-                    return displayName(node, ps);
-                case "emailaddress":
-                    return emailAddress(node, ps);
-                case "contactbirthday":
-                    return contactBirthday(node, ps);
-                default:
-                    return (AIMLProcessor.genericXML(node, ps));
-            }
+            return switch (nodeName) {
+                case "contactid" -> contactId(node, ps);
+                case "multipleids" -> multipleIds(node, ps);
+                case "dialnumber" -> dialNumber(node, ps);
+                case "addinfo" -> newContact(node, ps);
+                case "displayname" -> displayName(node, ps);
+                case "emailaddress" -> emailAddress(node, ps);
+                case "contactbirthday" -> contactBirthday(node, ps);
+                default -> (AIMLProcessor.genericXML(node, ps));
+            };
         } catch (Exception ex) {
-            ex.printStackTrace();
+            log.error("recursEval Error", ex);
             return "";
         }
     }
