@@ -7,7 +7,6 @@ package org.alicebot.ab.utils;
 
 import com.mayabot.nlp.fasttext.FastText;
 import com.mayabot.nlp.fasttext.ScoreLabelPair;
-import fasttext.FastTextPrediction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,18 +19,21 @@ import java.lang.reflect.Method;
 import java.net.URL;
 import java.net.URLClassLoader;
 import java.text.Normalizer;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 /**
  * Custom methods for polish language and jar plugin call. 
  * @author skost
  */
 public class SprintUtils {
+    private static Map<String, Class<?>> classCache = new HashMap<>();
     private static final Logger log = LoggerFactory.getLogger(SprintUtils.class);
-   
-    public static Map<String, fasttext.FastText> mlaModels; 
+
+    public static void main(String[] args) {
+        System.out.println(Locale.forLanguageTag("pl"));
+
+    }
+
     public static Map<String, FastText> mlModels;
     /**
      * Replace polish marks in string.
@@ -40,8 +42,7 @@ public class SprintUtils {
      * @return 
      */
     @Deprecated
-    public static String unaccent(String src, boolean isPolishMarks) 
-    {
+    public static String unaccent(String src, boolean isPolishMarks) {
         String temp = src;
         if(src==null)
             return null;
@@ -53,12 +54,10 @@ public class SprintUtils {
 
 
 
-    public static boolean updateMlModel(String model)
-    {
+    public static boolean updateMlModel(String model) {
 
         FastText fastText = mlModels.get(model);
-        if(fastText != null)
-        {
+        if(fastText != null) {
             log.info("updateMlModel. Model " + model + " removed from system.");
             mlModels.remove(model);
         }
@@ -67,17 +66,14 @@ public class SprintUtils {
 
         File file = new File(path);
 
-        if(!file.exists())
-        {
-            log.error("updateMlModel. Model " + model + " not exists. Path: " + file.getAbsolutePath());
+        if(!file.exists()) {
+            log.error("updateMlModel. Model {} not exists. Path: {}", model, file.getAbsolutePath());
             return false;
         }
         try {
             fastText = FastText.Companion.loadModelFromSingleFile(file);
-        }
-        catch (Exception ex)
-        {
-            log.error("updateMlModel ERROR : " + ex, ex);
+        } catch (Exception ex) {
+            log.error("updateMlModel ERROR : {}", ex, ex);
             return false;
         }
 
@@ -86,60 +82,7 @@ public class SprintUtils {
         return true;
     }
             
-    public static String mla(String model, String threshold, String score, String parameter, String sessionId)
-    {
 
-        String out;
-        int iMinScore = 0;
-        if(score != null && score.length() > 0)
-            iMinScore = Integer.parseInt(score);
-
-        float fThreshold = 0f;
-
-        if(threshold != null && threshold.length() > 0)
-            fThreshold = Float.parseFloat(threshold);
-
-        log.info("mla Request: sessionId: " + sessionId + " Model: " + model + " MinScore: " + iMinScore + " Threshold: " + fThreshold + " parameter: " + parameter);
-
-
-        try {
-
-            fasttext.FastText ftmodel = mlaModels.get(model);
-
-            if(ftmodel == null)
-            {
-                log.warn(sessionId + "\tInvalid model name");
-                return "ERR Invalid model name";
-            }
-
-            List<FastTextPrediction> result = ftmodel.predictAll(Arrays.asList(parameter.split(" ")),fThreshold);
-
-            int iScore = (int) (result.get(0).probability() * 100);
-
-            if (!result.get(0).label().equals("__label__oos"))
-            {
-                log.info(sessionId + "\tOK:\t" + parameter + "\tresult : " + result.get(0).label() + " score: " + iScore);
-            } else {
-                log.warn(sessionId + "\tNO QUALIFICATION:\t" + parameter);
-            }
-
-            if(iScore >= iMinScore)
-                out= result.get(0).label() + " " + iScore;
-            else
-                out = "__label__oos" + " " + iScore;
-
-
-
-            //out= result.get(0).label() + " " + iScore;
-
-        } catch (Exception e) {
-
-            out = "ERR " + e.getMessage();
-            log.error(sessionId + "\tpredictSupervisedModel ERROR : " + e, e);
-        }
-
-        return out;
-    }
 
     /**
      * Predict fastText label tranined suprvised model
@@ -151,36 +94,33 @@ public class SprintUtils {
      * @param sessionId Bot SessionId
      * @return
      */
-    public static String ml(String model, String nBest, String threshold, String score, String parameter, String sessionId)
-    {
+    public static String ml(String model, String nBest, String threshold, String score, String parameter, String sessionId) {
         String out = "ERR";
 
         try {
-            int iNbest = Integer.parseInt(nBest);
+            int best = Integer.parseInt(nBest);
             float fThreshold = Float.parseFloat(threshold);
             int iMinScore = Integer.parseInt(score);
 
 
-            log.info("ml Request: sessionId: " + sessionId + " Model: " + model + " Nbest: " + iNbest + " Threshold: " + fThreshold + " MinScore: " + iMinScore + " parameter: " + parameter);
+            log.info("ml Request: sessionId: {} Model: {} Nbest: {} Threshold: {} MinScore: {} parameter: {}", sessionId, model, best, fThreshold, iMinScore, parameter);
 
             FastText fastText = mlModels.get(model);
 
-            if(fastText == null)
-            {
-                log.warn(sessionId + "\tInvalid model name");
+            if(fastText == null) {
+                log.warn("{}\tML Invalid model name", sessionId);
                 return "ERR Invalid model name";
             }
 
-            List<ScoreLabelPair> result = fastText.predict(Arrays.asList(parameter.split(" ")), iNbest, fThreshold);
+            List<ScoreLabelPair> result = fastText.predict(Arrays.asList(parameter.split(" ")), best, fThreshold);
 
-            log.info("ml Response sessionId: " + sessionId + " result.size: " + result.size());
+            log.info("ml Response sessionId: {} result.size: {}", sessionId, result.size());
             for(ScoreLabelPair pair : result) {
                 int iScore = (int) (pair.getScore() * 100);
 
-                log.info("Response sessionId: " + sessionId + " parameter: " + parameter + " RESPONSE score: " + pair.getScore() + " iScore: " + iScore + " label: " + pair.getLabel() + " MinScore: " + iMinScore);
+                log.info("Response sessionId: {} parameter: {} RESPONSE score: {} iScore: {} label: {} MinScore: {}", sessionId, parameter, pair.getScore(), iScore, pair.getLabel(), iMinScore);
 
-                if(iScore >= iMinScore)
-                {
+                if(iScore >= iMinScore) {
                     out = pair.getLabel() + " " + iScore;
                     break;
                 }
@@ -191,12 +131,16 @@ public class SprintUtils {
         } catch (Exception e) {
 
             out = "ERR " + e.getMessage();
-            log.error("predictSupervisedModel ERROR : " + e, e);
+            log.error("predictSupervisedModel ERROR", e);
         }
 
         return out;
     }
-    
+
+
+    public static void resetClassCache() {
+        classCache.clear();
+    }
     
     /**
      * Java jar integration method. 
@@ -207,53 +151,39 @@ public class SprintUtils {
      * @param sessionId sessionid
      * @return plugin reponse
      */
-    public static String callPlugin(String file, String classLoad, String methodName, String parameter, String sessionId)
-    {
-        URLClassLoader urlClassLoader = null;
-        //String f = "jar:file:///" + file + "!/";
-        
+    public static String callPlugin(String file, String classLoad, String methodName, String parameter, String sessionId) {
         File f = new File(file);
-        
-        String out = "";
+        String out;
+
         try {
-            
-            urlClassLoader = new URLClassLoader(new URL[] {f.toURI().toURL()},
-                                         SprintUtils.class.getClassLoader());
-                                    
-            //URL[] classLoaderUrls = new URL[]{new URL(f)};         
-            // Create a new URLClassLoader 
-            //urlClassLoader = new URLClassLoader(classLoaderUrls);
+            Class<?> bean = classCache.get(classLoad);
+            if (bean == null) {
+                try (URLClassLoader urlClassLoader = new URLClassLoader(new URL[]{f.toURI().toURL()},
+                        SprintUtils.class.getClassLoader())) {
 
-            // Load the target class
-            Class<?> beanClass = urlClassLoader.loadClass(classLoad);
+                    // Load the target class
+                    Class<?> beanClass = urlClassLoader.loadClass(classLoad);
 
-            // Create a new instance from the loaded class
-            Constructor<?> constructor = beanClass.getConstructor();             
+                    classCache.put(classLoad, beanClass);
 
-            Object beanObj = constructor.newInstance();        
-            // Getting a method from the loaded class and invoke it
-                        
-            Method method = beanClass.getMethod("processCustomResultPocessor", String.class, String.class, String.class);    
-            String response = (String) method.invoke(beanObj, sessionId, parameter, methodName);
-            log.info("Request: sessionId: " + sessionId + " parameter: " + parameter + " method: " + methodName + " plugin response: " + response);
-            
-            out = response;                                    
-        } 
-        catch (Exception e) 
-        {
-            out = "ERR " + e.getMessage();
-            log.error("callPlugin file: " + f + " parameter : " +parameter + " ERROR : " + e, e);
+                    Constructor<?> constructor = beanClass.getConstructor();
+                    Object beanObj = constructor.newInstance();
+                    Method method = beanClass.getMethod("processCustomResultPocessor", String.class, String.class, String.class);
+                    out = (String) method.invoke(beanObj, sessionId, parameter, methodName);
+
+                }
+            } else {
+                Constructor<?> constructor = bean.getConstructor();
+                Object beanObj = constructor.newInstance();
+                Method method = bean.getMethod("processCustomResultPocessor", String.class, String.class, String.class);
+                out = (String) method.invoke(beanObj, sessionId, parameter, methodName);
+                log.info("call from loaded class: {}", classLoad);
+            }
+        } catch (Exception ex) {
+            out = "ERR " + ex.getMessage();
+            log.error("callPlugin file: {} parameter : {} ERROR", f, parameter, ex);
         }
-        finally {
-            if(urlClassLoader != null)
-                try {
-                    urlClassLoader.close();
-                } catch (IOException ex) {
-                    out = "ERR " + ex.getMessage();
-                    log.error("callPlugin urlClassLoader.close() file: " + f + " parameter : " +parameter + " ERROR : " + ex, ex);
-                }                       
-        } 
-        
+        log.info("{} : request parameter: {} method: {} plugin response: {}", sessionId, parameter, methodName, out);
         return out;
     }
     
@@ -265,8 +195,7 @@ public class SprintUtils {
      * @return 
      */
     @Deprecated
-    public static String readBashScript(String scrip, String parameters)
-    {                
+    public static String readBashScript(String scrip, String parameters) {
         String out = "";
         
         if (scrip == null)
@@ -284,13 +213,13 @@ public class SprintUtils {
             try {
                 proc.waitFor();
             } catch (InterruptedException ex) {
-                log.error("readBashScript script: " + scrip + " parameters : " +parameters + " InterruptedException ERROR : " + ex, ex); 
+                log.error("readBashScript script: {} parameters : {} InterruptedException ERROR : {}", scrip, parameters, ex, ex);
             }
             while (read.ready()) {
                 out = read.readLine();
             }
         } catch (IOException e) {
-            log.error("readBashScript script: " + scrip + " parameters : " +parameters + " IOException ERROR : " + e, e); 
+            log.error("readBashScript script: {} parameters : {} IOException ERROR : {}", scrip, parameters, e, e);
         }
         
         log.info("readBashScript: scrip = {} parameters = {} [response = {}]",scrip, parameters, out);
